SUMMARY:  added static cast in Teuchos::OrdinalTraits<size_t>
     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  n/a

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  n/a

     3)  Did you modify or add any source or build system files?  yes
     3a) Did you rebuild the libraries and executables?  yes
     3b) Did you rerun the test suite on the packages that you changed and
         all packages that depend on the modified packages?  yes
     3c) Did you change the required or optional dependencies for any
         package?  n/a


     DESCRIPTION OF CHANGES:

OrdinalTraits<size_t>::invalid() returned -1 without static_cast<size_t> on systems with a signed size_t; this did not come up because my test machines use unsigned size_t.


(checklist version 2009-07-14)

+++ packages/teuchos/src/Teuchos_OrdinalTraits.hpp
@@ -136,7 +136,7 @@ struct OrdinalTraits<size_t> {
   static const bool hasMachineParameters = false;
   static inline size_t zero()              {return(static_cast<size_t>(0));}
   static inline size_t one()               {return(static_cast<size_t>(1));}
-  static inline size_t invalid()           {return(std::numeric_limits<size_t>::is_signed ? -1 : std::numeric_limits<size_t>::max());}
+  static inline size_t invalid()           {return(std::numeric_limits<size_t>::is_signed ? static_cast<size_t>(-1)            : std::numeric_limits<size_t>::max()  );}
   static inline size_t max()               {return(std::numeric_limits<size_t>::is_signed ? std::numeric_limits<size_t>::max() : std::numeric_limits<size_t>::max()-1);}
   static inline std::string name()         {return("size_t");}
 };
